+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # HEAD
 
 - Adding: attr_reader for `:path, :html_options, :width, :height`
+- Adding: `keywords.json` and `codepoints.json` methods to the octicon object. Making them available via the api
 
 # 1.1.0